Create README.md
In this work, The project constructed multiple ensemble machine-learning models based on S-AKI patient data from the MIMIC-IV repository. The best model, XGBoost, was ultimately selected to provide early warning of the 28-day mortality risk in S-AKI cases. The prediction models were interpreted and visualized by Shap values. Visualization of the critical variables associated with the risk of death in individual cases was performed with individualized force plots. In particular, we applied COX regression-restricted cubic spline plots to obtain nonlinear relationships between nine key continuous independent variables and 28-day mortality. Our model is to inform treatment decisions for patients with S-AKI. It provides a basis for stratified and graded management of patients and improves survival rates.